Avoid evaluating (INT_MIN/-1) with 32 bit signed numbers. It can cause SIGFPE.

+++ src/vdbe.c
@@ -1171,9 +1171,9 @@ case OP_Remainder: {           /* same as TK_REM, no-push */
         break;
       }
       default: {
-        int ia = (int)a;
-        int ib = (int)b;
-        if( ia==0.0 ) goto divide_by_zero;
+        i64 ia = (i64)a;
+        i64 ib = (i64)b;
+        if( ia==0 ) goto divide_by_zero;
         b = ib % ia;
         break;
       }
